+++ spectrochempy/utils/file.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # ======================================================================================
-# Copyright (©) 2015-2023 LCS - Laboratoire Catalyse et Spectrochimie, Caen, France.
+# Copyright (©) 2015-2025 LCS - Laboratoire Catalyse et Spectrochimie, Caen, France.
 # CeCILL-B FREE SOFTWARE LICENSE AGREEMENT
 # See full LICENSE agreement in the root directory.
 # ======================================================================================
@@ -478,7 +478,7 @@ def get_filenames(*filenames, **kwargs):
 
         getdir = kwargs.get(
             "iterdir",
-            directory is not None or kwargs.get("protocol", None) == ["topspin"]
+            directory is not None or kwargs.get("protocol", None) == ["topspin"],
             # or kwargs.get("protocol", None) == ["carroucell"],
         )