+++ tests/test_core/test_readers/test_importer.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # ======================================================================================
-# Copyright (©) 2015-2023 LCS - Laboratoire Catalyse et Spectrochimie, Caen, France.
+# Copyright (©) 2015-2025 LCS - Laboratoire Catalyse et Spectrochimie, Caen, France.
 # CeCILL-B FREE SOFTWARE LICENSE AGREEMENT
 # See full LICENSE agreement in the root directory.
 # ======================================================================================
@@ -26,6 +26,7 @@ from spectrochempy.core.readers.importer import (
 from spectrochempy.utils.file import pathclean
 
 DATADIR = prefs.datadir
+
 
 # Simulation of a read function
 def read_fake(*paths, **kwargs):