+++ tests/test_utils/test_print_versions.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # ======================================================================================
-# Copyright (©) 2015-2023 LCS - Laboratoire Catalyse et Spectrochimie, Caen, France.
+# Copyright (©) 2015-2025 LCS - Laboratoire Catalyse et Spectrochimie, Caen, France.
 # CeCILL-B FREE SOFTWARE LICENSE AGREEMENT
 # See full LICENSE agreement in the root directory.
 # ======================================================================================
@@ -13,5 +13,10 @@ import spectrochempy as scp
 def test_show_versions() -> None:
     f = io.StringIO()
     scp.show_versions(file=f)
-    assert "INSTALLED VERSIONS" in f.getvalue()
-    assert "python:" in f.getvalue()
+    assert "INSTALLED PACKAGES" in f.getvalue()
+    assert "DEV PACKAGES" in f.getvalue()
+    assert "- pytest" in f.getvalue()
+    assert "- python" in f.getvalue()
+    assert "- numpy" in f.getvalue()
+    assert "- pint" in f.getvalue()
+    assert "SPECTROCHEMPY" in f.getvalue()